@@ -16,8 +16,8 @@
 qclib auxiliary functions
 """
 
-from qiskit import execute, transpile
-from qiskit.providers.aer.backends import AerSimulator
+from qiskit import transpile
+from qiskit_aer import AerSimulator
 import numpy as np
 from scipy import sparse
 
@@ -33,7 +33,10 @@ def get_counts(circ):
     counts: output generated by the quantum circuit
     """
     backend = AerSimulator()
-    counts = execute(circ, backend).result().get_counts()
+    counts = backend.run(
+        transpile(circ, backend)
+    ).result().get_counts()
+
     return counts
 
 
@@ -54,6 +57,42 @@ def get_state(circ):
     state_vector = backend.run(tcirc).result().get_statevector()
 
     return np.array(state_vector)
+
+
+def get_cnot_count(circ, optimization_level=0):
+    """
+    Parameters
+    ----------
+    circ: QuantumCircuit
+
+    Returns
+    -------
+    cnot_count: number of cnot gates in the quantum circuit
+    """
+    tcirc = transpile(
+        circ,
+        basis_gates=['u', 'cx'],
+        optimization_level=optimization_level
+    )
+    return tcirc.count_ops().get('cx', 0)
+
+
+def get_depth(circ, optimization_level=0):
+    """
+    Parameters
+    ----------
+    circ: QuantumCircuit
+
+    Returns
+    -------
+    circuit depth: depth of the quantum circuit
+    """
+    tcirc = transpile(
+        circ,
+        basis_gates=['u', 'cx'],
+        optimization_level=optimization_level
+    )
+    return tcirc.depth()
 
 
 def replace_all_values_with(new_value, dataset):
@@ -168,7 +207,7 @@ def _double_sparse_binary(nbits, log_size, p_1, p_0):
     return bin_data
 
 
-def double_sparse(nbits, log_size, p_1):
+def double_sparse(nbits, log_size, p_1, complex_amplitudes=True):
     """
     Parameters
     ----------
@@ -180,15 +219,18 @@ def double_sparse(nbits, log_size, p_1):
     -------
     \\sum_{k} x_k |p_k>, each bit of p_k is equal to 1 with probability p1
     """
-    data = np.random.rand(2**log_size) + np.random.rand(2**log_size) * 1j
+    if complex_amplitudes:
+        data = np.random.rand(2**log_size) + np.random.rand(2**log_size) * 1j
+    else:
+        data = np.random.rand(2**log_size)
     length = np.linalg.norm(data)
     data = (1 / length) * data
 
     binary = _double_sparse_binary(nbits, log_size, 1 - p_1, p_1)
     bin_data = [(binary[i], data[i]) for i in range(2**log_size)]
-
     bin_data.sort(key=_count_ones)
-    return bin_data
+
+    return {''.join(map(str, b)):d for b, d in bin_data}
 
 
 def _compute_matrix_angles(feature, norm):
@@ -235,10 +277,10 @@ def build_state_dict(state):
     Args:
       state: The classical description of the state vector
     """
-    n_qubits = np.ceil(np.log(len(state))).astype(int)
+    n_qubits = np.ceil(np.log2(len(state))).astype(int)
     state_dict = {}
     for (value_idx, value) in enumerate(state):
         if value != 0:
-            binary_string = "{:0{}b}".format(value_idx, n_qubits)[::-1]
+            binary_string = f"{value_idx:0{n_qubits}b}"[::-1]
             state_dict[binary_string] = value
     return state_dict
