@@ -17,8 +17,12 @@
 import numpy as np
 from qiskit import QuantumCircuit, QuantumRegister
 from qiskit.circuit.library.standard_gates import UGate
+from qiskit.quantum_info import Operator
 from qclib.util import _compute_matrix_angles
 from qclib.gates.initialize_sparse import InitializeSparse
+from qclib.gates.mcg import Mcg
+from qclib.gates.ldmcsu import LdMcSpecialUnitary
+
 
 # pylint: disable=maybe-no-member
 
@@ -41,11 +45,17 @@ class CvoqramInitialize(InitializeSparse):
         default_with_aux = True
         if opt_params is None:
             self.with_aux = default_with_aux
+            self.mcg_method = 'linear'
         else:
             if opt_params.get("with_aux") is None:
                 self.with_aux = default_with_aux
             else:
                 self.with_aux = opt_params.get("with_aux")
+
+            if opt_params.get("mcg_method") is None:
+                self.mcg_method = 'linear'
+            else:
+                self.mcg_method = opt_params.get("mcg_method")
 
         if label is None:
             label = "CVOSP"
@@ -137,8 +147,15 @@ class CvoqramInitialize(InitializeSparse):
             if self.with_aux:
                 self._mcuvchain(circuit, theta, phi, lam)
             else:
-                gate = UGate(theta, phi, lam).control(len(control))
-                circuit.append(gate, memory[control] + [self.aux[0]])
+                gate = UGate(theta, phi, lam)
+                if self.mcg_method == 'qiskit':
+                    circuit.append(gate.control(len(control)), memory[control] + [self.aux[0]])
+                elif self.mcg_method == 'barenco':
+                    gate_op = Operator(gate).data
+                    LdMcSpecialUnitary.ldmcsu(circuit, gate_op, memory[control], [self.aux[0]])
+                else:
+                    gate_op = Operator(gate).data
+                    Mcg.mcg(circuit, gate_op, memory[control], [self.aux[0]])
 
         self.norm = self.norm - np.absolute(np.power(feature, 2))
 
