@@ -19,7 +19,7 @@
 import numpy as np
 from qiskit import QuantumCircuit, QuantumRegister
 from qiskit.circuit.library import UGate
-from qclib.gates.mc_gate import mc_gate
+from qclib.gates.ldmcu import Ldmcu
 from qclib.gates.initialize_sparse import InitializeSparse
 
 
@@ -176,10 +176,6 @@ class MergeInitialize(InitializeSparse):
             else:
                 dif_values.append("1")
                 temp_strings = t_1
-
-            # dif_qubits must have at least two values stored in it
-            if (len(temp_strings) - 1) == 1 and len(dif_qubits) == 1:
-                temp_strings = b_strings
 
         return temp_strings, dif_qubits, dif_values
 
@@ -444,7 +440,7 @@ class MergeInitialize(InitializeSparse):
             quantum_circuit.append(merge_gate, dif_qubits + [dif], [])
         else:
             gate_definition = UGate(theta, phi, lamb, label="U").to_matrix()
-            mc_gate(gate_definition, quantum_circuit, dif_qubits, dif)
+            Ldmcu.ldmcu(quantum_circuit, gate_definition, dif_qubits, dif)
 
         state_dict = self._update_state_dict_according_to_operation(
             state_dict, "merge", None, merge_strings=[bitstr1, bitstr2]
