@@ -16,7 +16,6 @@
 https://arxiv.org/abs/2108.10182
 """
 
-import cmath
 import math
 from dataclasses import dataclass
 from qclib.state_preparation.util.tree_utils import is_leaf
@@ -36,14 +35,12 @@ class NodeAngleTree:
     right: "NodeAngleTree"
 
     def __str__(self):
-        return (
-            str(self.level)
-            + "_"
-            + str(self.index)
-            + "\n"
-            + f"{self.angle_y:.2g}_"
-            + f"{self.angle_z:.2g}"
-        )
+        space = '\t' * self.level
+        txt = f"{space * self.level} y {self.angle_y:.2f} z{self.angle_z:.2f}\n"
+        if self.left is not None:
+            txt += self.left.__str__()
+            txt += self.right.__str__()
+        return txt
 
 
 def create_angles_tree(state_tree):
@@ -52,19 +49,21 @@ def create_angles_tree(state_tree):
     :param tree: used in the recursive calls
     :return: tree with angles that will be used to perform the state preparation
     """
-    amp = 0.0
-    if state_tree.amplitude != 0.0:
-        amp = state_tree.right.amplitude / state_tree.amplitude
+    mag = 0.0
+    if state_tree.mag != 0.0:
+        mag = state_tree.right.mag / state_tree.mag
+
+    arg = state_tree.right.arg - state_tree.arg
 
     # Avoid out-of-domain value due to numerical error.
-    if abs(amp) < -1.0:
+    if mag < -1.0:
         angle_y = -math.pi
-    elif abs(amp) > 1.0:
+    elif mag > 1.0:
         angle_y = math.pi
     else:
-        angle_y = 2 * math.asin(abs(amp))
+        angle_y = 2 * math.asin(mag)
 
-    angle_z = 2 * cmath.phase(amp)
+    angle_z = 2 * arg
 
     node = NodeAngleTree(
         state_tree.index, state_tree.level, angle_y, angle_z, None, None
