@@ -17,7 +17,7 @@
 from unittest import TestCase
 import numpy as np
 from qiskit import ClassicalRegister
-from qiskit.providers.aer.backends import AerSimulator
+from qiskit_aer import AerSimulator
 from qclib.state_preparation import BdspInitialize
 from .util import measurement
 
